+++ crates/polars-mem-engine/src/executors/scan/ipc.rs
@@ -29,7 +29,7 @@ impl IpcExec {
         };
         let force_async = config::force_async();
 
-        let mut out = if is_cloud || (self.sources.is_files() && force_async) {
+        let mut out = if is_cloud || (self.sources.is_paths() && force_async) {
             feature_gated!("cloud", {
                 if force_async && config::verbose() {
                     eprintln!("ASYNC READING FORCED");
@@ -102,7 +102,7 @@ impl IpcExec {
                     self.file_options
                         .include_file_paths
                         .as_ref()
-                        .map(|x| (x.clone(), Arc::from(source.to_file_path()))),
+                        .map(|x| (x.clone(), Arc::from(source.to_include_path_name()))),
                 )
                 .finish()
         };
