+++ crates/polars-plan/src/plans/ir/scan_sources.rs
@@ -1,0 +1,270 @@
+use std::fs::File;
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+
+use polars_core::error::{feature_gated, PolarsResult};
+use polars_utils::mmap::MemSlice;
+use polars_utils::pl_str::PlSmallStr;
+
+use super::DslScanSources;
+
+/// Set of sources to scan from
+///
+/// This is can either be a list of paths to files, opened files or in-memory buffers. Mixing of
+/// buffers is not currently possible.
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Debug, Clone)]
+pub enum ScanSources {
+    Paths(Arc<[PathBuf]>),
+
+    #[cfg_attr(feature = "serde", serde(skip))]
+    Files(Arc<[File]>),
+    #[cfg_attr(feature = "serde", serde(skip))]
+    Buffers(Arc<[bytes::Bytes]>),
+}
+
+/// A reference to a single item in [`ScanSources`]
+#[derive(Debug, Clone, Copy)]
+pub enum ScanSourceRef<'a> {
+    Path(&'a Path),
+    File(&'a File),
+    Buffer(&'a bytes::Bytes),
+}
+
+/// An iterator for [`ScanSources`]
+pub struct ScanSourceIter<'a> {
+    sources: &'a ScanSources,
+    offset: usize,
+}
+
+impl Default for ScanSources {
+    fn default() -> Self {
+        Self::Buffers(Arc::default())
+    }
+}
+
+impl std::hash::Hash for ScanSources {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        std::mem::discriminant(self).hash(state);
+
+        // @NOTE: This is a bit crazy
+        //
+        // We don't really want to hash the file descriptors or the whole buffers so for now we
+        // just settle with the fact that the memory behind Arc's does not really move. Therefore,
+        // we can just hash the pointer.
+        match self {
+            Self::Paths(paths) => paths.hash(state),
+            Self::Files(files) => files.as_ptr().hash(state),
+            Self::Buffers(buffers) => buffers.as_ptr().hash(state),
+        }
+    }
+}
+
+impl PartialEq for ScanSources {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (ScanSources::Paths(l), ScanSources::Paths(r)) => l == r,
+            (ScanSources::Files(l), ScanSources::Files(r)) => std::ptr::eq(l.as_ptr(), r.as_ptr()),
+            (ScanSources::Buffers(l), ScanSources::Buffers(r)) => {
+                std::ptr::eq(l.as_ptr(), r.as_ptr())
+            },
+            _ => false,
+        }
+    }
+}
+
+impl Eq for ScanSources {}
+
+impl ScanSources {
+    pub fn iter(&self) -> ScanSourceIter {
+        ScanSourceIter {
+            sources: self,
+            offset: 0,
+        }
+    }
+
+    pub fn to_dsl(self, is_expanded: bool) -> DslScanSources {
+        DslScanSources {
+            sources: self,
+            is_expanded,
+        }
+    }
+
+    /// Are the sources all paths?
+    pub fn is_paths(&self) -> bool {
+        matches!(self, Self::Paths(_))
+    }
+
+    /// Try cast the scan sources to [`ScanSources::Paths`]
+    pub fn as_paths(&self) -> Option<&[PathBuf]> {
+        match self {
+            Self::Paths(paths) => Some(paths.as_ref()),
+            Self::Files(_) | Self::Buffers(_) => None,
+        }
+    }
+
+    /// Try cast the scan sources to [`ScanSources::Paths`] with a clone
+    pub fn into_paths(&self) -> Option<Arc<[PathBuf]>> {
+        match self {
+            Self::Paths(paths) => Some(paths.clone()),
+            Self::Files(_) | Self::Buffers(_) => None,
+        }
+    }
+
+    /// Try get the first path in the scan sources
+    pub fn first_path(&self) -> Option<&Path> {
+        match self {
+            Self::Paths(paths) => paths.first().map(|p| p.as_path()),
+            Self::Files(_) | Self::Buffers(_) => None,
+        }
+    }
+
+    /// Is the first path a cloud URL?
+    pub fn is_cloud_url(&self) -> bool {
+        self.first_path().is_some_and(polars_io::is_cloud_url)
+    }
+
+    pub fn len(&self) -> usize {
+        match self {
+            Self::Paths(s) => s.len(),
+            Self::Files(s) => s.len(),
+            Self::Buffers(s) => s.len(),
+        }
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    pub fn first(&self) -> Option<ScanSourceRef> {
+        self.get(0)
+    }
+
+    /// Turn the [`ScanSources`] into some kind of identifier
+    pub fn id(&self) -> PlSmallStr {
+        if self.is_empty() {
+            return PlSmallStr::from_static("EMPTY");
+        }
+
+        match self {
+            Self::Paths(paths) => {
+                PlSmallStr::from_str(paths.first().unwrap().to_string_lossy().as_ref())
+            },
+            Self::Files(_) => PlSmallStr::from_static("OPEN_FILES"),
+            Self::Buffers(_) => PlSmallStr::from_static("IN_MEMORY"),
+        }
+    }
+
+    /// Get the scan source at specific address
+    pub fn get(&self, idx: usize) -> Option<ScanSourceRef> {
+        match self {
+            Self::Paths(paths) => paths.get(idx).map(|p| ScanSourceRef::Path(p)),
+            Self::Files(files) => files.get(idx).map(ScanSourceRef::File),
+            Self::Buffers(buffers) => buffers.get(idx).map(ScanSourceRef::Buffer),
+        }
+    }
+
+    /// Get the scan source at specific address
+    ///
+    /// # Panics
+    ///
+    /// If the `idx` is out of range.
+    #[track_caller]
+    pub fn at(&self, idx: usize) -> ScanSourceRef {
+        self.get(idx).unwrap()
+    }
+}
+
+impl<'a> ScanSourceRef<'a> {
+    /// Get the name for `include_paths`
+    pub fn to_include_path_name(&self) -> &str {
+        match self {
+            Self::Path(path) => path.to_str().unwrap(),
+            Self::File(_) => "open-file",
+            Self::Buffer(_) => "in-mem",
+        }
+    }
+
+    /// Turn the scan source into a memory slice
+    pub fn to_memslice(&self) -> PolarsResult<MemSlice> {
+        self.to_memslice_possibly_async(false, None, 0)
+    }
+
+    pub fn to_memslice_async_latest(&self, run_async: bool) -> PolarsResult<MemSlice> {
+        match self {
+            ScanSourceRef::Path(path) => {
+                let file = if run_async {
+                    feature_gated!("cloud", {
+                        polars_io::file_cache::FILE_CACHE
+                            .get_entry(path.to_str().unwrap())
+                            // Safety: This was initialized by schema inference.
+                            .unwrap()
+                            .try_open_assume_latest()?
+                    })
+                } else {
+                    polars_utils::open_file(path)?
+                };
+
+                Ok(MemSlice::from_mmap(Arc::new(unsafe {
+                    memmap::Mmap::map(&file)?
+                })))
+            },
+            ScanSourceRef::File(file) => Ok(MemSlice::from_mmap(Arc::new(unsafe {
+                memmap::Mmap::map(*file)?
+            }))),
+            ScanSourceRef::Buffer(buff) => Ok(MemSlice::from_bytes((*buff).clone())),
+        }
+    }
+
+    pub fn to_memslice_possibly_async(
+        &self,
+        run_async: bool,
+        #[cfg(feature = "cloud")] cache_entries: Option<
+            &Vec<Arc<polars_io::file_cache::FileCacheEntry>>,
+        >,
+        #[cfg(not(feature = "cloud"))] cache_entries: Option<&()>,
+        index: usize,
+    ) -> PolarsResult<MemSlice> {
+        match self {
+            Self::Path(path) => {
+                let f = if run_async {
+                    feature_gated!("cloud", {
+                        cache_entries.unwrap()[index].try_open_check_latest()?
+                    })
+                } else {
+                    polars_utils::open_file(path)?
+                };
+
+                let mmap = unsafe { memmap::Mmap::map(&f)? };
+                Ok(MemSlice::from_mmap(Arc::new(mmap)))
+            },
+            Self::File(file) => {
+                let mmap = unsafe { memmap::Mmap::map(*file)? };
+                Ok(MemSlice::from_mmap(Arc::new(mmap)))
+            },
+            Self::Buffer(buff) => Ok(MemSlice::from_bytes((*buff).clone())),
+        }
+    }
+}
+
+impl<'a> Iterator for ScanSourceIter<'a> {
+    type Item = ScanSourceRef<'a>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let item = match self.sources {
+            ScanSources::Paths(paths) => ScanSourceRef::Path(paths.get(self.offset)?),
+            ScanSources::Files(files) => ScanSourceRef::File(files.get(self.offset)?),
+            ScanSources::Buffers(buffers) => ScanSourceRef::Buffer(buffers.get(self.offset)?),
+        };
+
+        self.offset += 1;
+        Some(item)
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let len = self.sources.len() - self.offset;
+        (len, Some(len))
+    }
+}
+
+impl<'a> ExactSizeIterator for ScanSourceIter<'a> {}
